+++ energy_onshore/solar_test.py
@@ -1,8 +1,0 @@
-# Pending to be completed.
-# def solar_capacity_factor_hourly(rsds, t, elevation):
-#    cf = xr.DataArray(
-#        cf, coords=coords, dims=dims, attrs=attrs, name=attrs["shortname"]
-#    )
-#
-#    pass
-